test: add question complex service coverage
